+++ androidApp/src/main/java/com/devjeong/watermelon_player/android/player/presentation/components/CoverScreen.kt
@@ -1,0 +1,2 @@
+package com.devjeong.watermelon_player.android.player.presentation.components
+
